Keep artifact commands in their script module
